requirements.txt
numpy
pandas
scikit-learn
Pymysql
python-dotenv